Add stub_libs properties to art-bootclasspath-fragment
This is being done as part of the hidden API modularization work. The
hidden API processing needs access to the stubs in order to make the
relevant members in the dex implementation jars accessible through the
APIs.

The jacoco-stubs are added to coverage.api.stub_libs as they need to be
added to the stubs when jacocoagent is added to the contents.

Bug: 179354495
Test: m art-module-sdk
      - make sure that the stubs are listed in the snapshot
Change-Id: I31a24a3cc99a862a90459aeae9a0b7a9c3e9fc40