Restructured build targets to be clearer
This CL makes service-art depend on a new target,
frameworks-module-common-unbundled-defaults, which contains all of the
settings that used to be in frameworks-module-common-defaults except for
the frameworks-annotations library.  This allows the ART module to use
the target while any issues with the annotation library are resolve,
thus simplifying the build configuration.

Test: m
Bug: 177273468
Change-Id: I12b2fea0c7abf64a8978d4debca5aa93228bfb9b