Fall back to the system namespace for loading libraries.
Using dlopen(3) in libnativeloader means loading from the
com_android_art namespace, because that's where libnativeloader lives.
However, since libnativeloader is loading libraries on behalf of other
callers that are external from ART, that's not appropriate to use - it
should only be used when loading libraries for ART's own purposes.

Hence, whenever we cannot find a better namespace through the
classloaders or other means, fall back to the system namespace. That
allows platform framework code to continue loading platform libraries
as usual, and is compatible with the old pre-namespace behavior.

Test: atest art/libnativeloader
Test: atest CtsJniTestCases
Bug: 130340935
Change-Id: I2513472c4e1c44f0fcb01a6b4f6eccbb03718f1d