Use only half of the cores in testrunner by default on target.
This is what we currently do on buildbots to avoid timeouts.
Set the appropriate default here and let the buildbots use the default.

This will make it device independent so we don't have to hard-code the
value on buildbot (in case device with different core count is used).

It will also make it easier to investigate the timeouts in the future,
hopefully allowing us fully utilize the devices on the build bots.

Allow the user to specify -j0 to explicitly use the default.

Bug: 142039427
Test: test.py on target
Change-Id: Ic7018c1a3239db25de148543bead62db16d02829