Update patch set 3
Patch Set 3:

Change bypassed Presubmit.

Bypass Reason: other 
Additional Feedback: The failing builds are already failing in ToT. If this was going to break a build it would break all the builds that include this repo but the other builds with the same manifest work.

Patch-set: 3
Tag: autogenerated:TreeHugger:Bypassed-Presubmit