ART: Redundant AND operation removal optimization
In the following code found in
SPECjvm2008 spec.benchmarks.mpegaudio.Harness.updateCRC32:
	b[i + length] = (byte) ((value & 0xff00) >> 8);
The AND operation is redundant because the lowest 8 bits will be
removed by the 8 bits right shift.

This optimization covers cases where a subsequent SHR and
Type Conversion (from Int and Short --> Byte make an AND redundant.
In such cases, the AND operation is removed. Patch also includes tests.
This patch covers only signed types.
Unsigned types will be supported in a future patch.

This optimization brings 3.3% perf increase on FFTBench workload.

Test: art/test.py -v --host --64 --run-test --optimizing -t\
458-checker-instruct-simplification

Change-Id: I948276eb59f92bbe04ae719002a5bd681af35260