Retry failed concurrent GC
If a concurrent GC fails to advance the GC number, we got stuck,
because that GC number was requested already, and hence no further
task was created. Yet this case appears possible, though unlikely.

When this happens, schedule a new task, rather than dropping the
request on the floor.

Add some comments and a DCHECK.

Bug: 186592536
Bug: 189150802

Test: Build and boot AOSP
Change-Id: Ib11734a9c87b9f9e19c5a3557eac9024f84cadf3