Add a aggregation count section to profile.
This section is used on server side.
The format is defined on server and it is:

***************
uint16 profile_aggregation_count
uint16 aggregation_count_for_each_dex_file_size
aggregation_count_for_each_dex_file[]

The format of aggregation_count_for_each_dex_file is:
  uint16 profile_index
  uint16 class_aggregation_count_size
  uint16 class_aggregation_count[]
  uint16 method_aggregation_count_size
  uint16 method_aggregation_count[]
***************

The aggregation count for classes/methods are in the ascending order of their indices.

Bug:
Test: m
Change-Id: I683b296699e2a80a8c6dc167c84b4d592a0f26fc