Binary search for method by name string and signature.
Start with a binary search for the name string and when
successful, compare the signature. On mismatch, repeat the
search with the found name index instead of the name string
until we fail to find the name, have a signature match or
end up with a method range with the same name index and
finish with a simple binary search for the signature.

Make this search sufficiently generic so that the code can
be easily reused for fields in the future.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Change-Id: I19b4779c019e9991b6d02ace57e378c029ba1174