Don't delete everything in apexdata/com.android.art.
Rather than deleting everything we now specifically delete the
artifacts directory when validation fails or forcing compilation. This
allows the pending directory, with CompOS artifacts, to be preserved
and used as an alternative.

As a slightly gratuitous bonus, delete the staging directory when
we're done with it, rather than just its contents.

Test: Manual, create pending directory, see it preserved.
Test: Presubmits
Bug: 190166662
Change-Id: Ic6f0cb15eb22b3a235d9df90653c6e6d63ea80bc