Copy host dex jars from the appropriate apex contents for the target
Previously, the host dex jars were copied from the <x>-hostdex modules
create in Make by the <x> module in Soong. Unfortunately, they did not
have hidden API flags encoded.

This change copies them from the appropriate apex's contents for the
target. That ensures that they will always be identical to that used
on a device. It does the same for conscrypt and core-icu4j so that it
will use the dex jars that would be installed on the device when
building from prebuilts.

Bug: 190491023
Test: lunch aosp_cf_x86_64_phone-userdebug
      rm -fr out/host/linux-x86/apex/
      m build-art-host-tests
      for i in $(cd out/host/linux-x86/apex/; find * -name \*.jar); do diff out/host/linux-x86/apex/$i out/target/product/vsoc_x86_64/apex/$i; done
Change-Id: Idfce8050c485a4658f1dcb1f7814505d5e0772a0