Fix OnDeviceSigningHostTest to correctly get mapped artifacts.
We use `pidof` to get the pid of a process, in order to get the mapped
files of the process from `/proc/<pid>/maps`. However, `pidof` may
output multiple pids in some cases, separated by spaces, which leads to
weird errors because the test cannot handle such output. After this
change, only the first pid is taken.

Bug: 189467174
Test: atest odsign_e2e_tests
Change-Id: I6e137819fa032611636f92fdd2b6c9faf58acd25