Opt to reduce gc load, reduce gc in some scenarios
1.Background GC,According to the ratio of the cpu running time and
duration time of this GC to represent the load of the system,
increase Multiplier appropriately when screen is on.

2.NativeAlloc GC,Increase the count of the number of small native
memory allocations.

3.CollectTransition GC,For process with a small heap size,recycling
and compress have no effect,so we can cancel it.

4.Explicit GC,Maybe we can consider changing Explicit GC to
concurrent GC,run in the heaptaskdaemon and won't block work thread anymore.
In addition, consider limiting the number of gc() calls
and the necessity of Explicit GC.
this CL is not incluede.

Verify 123 program, the equipment(oppo findx3pro) use time
can be extended by 18 minutes,and can effectively reduce the GC load.
The internal trial users have been tested for 3 weeks,
and no stability and compatibility issues have occurred.

Colleagues in charge of ART, please evaluate this CL
and hope to incorporate and release it in the Android S version.

Bug:200116730
Change-Id: Id2977f05eb249691326955e6f2424d4e5e08b417