Reland: Implement another OdrDexopt to use new CompOS API
The old OdrDexoptCompilationOS (now renamed) is still active (currently
controlled by a constant in the code) until everything is in place.

The new OdrDexoptCompilationOS put DexoptBcpExtArgs or
DexoptSystemServerArgs into ExtendableParcelable with TaskType.
The ExtendableParcelable and TaskType makes it easy for ART to add other
task types to run in CompOS.

Bug: 193668901
Test: atest ComposHostTestCases (with other changes)
Test: cd art/odrefresh; mma
Test: atest art_standalone_odrefresh_tests

Change-Id: I371e87e87197a5e2ce4ba8a73a1ba2add388a441