Provide API for CompOS to reconstruct dexopt context
The public API exposed by the ART APEX is meant to be used by CompOS
APEX as a client. In order to let ART maintain the flexibility to evolve
the compiler arguments (e.g. more/less flags/files), the API argument is
an opaque object that ART itself marshalled. On the API call, ART can
unmarshal the request since they are built in the same package, even
though the marshalled request may be routed to outside (in this case, to
run in the CompOS VM).

The API constructs the dex2oat command line arguments instead of
executing the command directly. This makes it the easiest to the client
to coordinate the fork(2). For example, when the client is a
multithreaded program, it is dangerous to fork(2) to execute the program
without involving the client. The alternative to make the API exec is
still possible though, if we expect the clients to get itself ready for
the exec.

Bug: 193668901
Test: atest ComposHostTestCases

Change-Id: I2199b4dcbd1d7409673d77660fc5bff5906abe5b