Use java_genrule for art run test data
The art run test data rules used genrule, but depended on a java_library
in tool_files.  For a host-only build like the aosp-build-tools branch
the java_library is disabled due to no device architecture enabled, but
the genrules were still enabled.  Use a java_genrule so they will
be disabled when no device architecture is enabled.

Bug: 202547639
Test: prebuilts/build-tools/build-prebuilts.sh
Change-Id: I1aa256106b525552a534547bd990d825abf71c86