Initialize the SafeBuffer to ensure the profile data consistent.
In profile v15, method_bitmap record in methods section is align to
8 bits(byte), so there may be some extra bits only for alignment(eg,
when 50 methods recorded, 54 bits used). These extra bits is not
initialized explicitly, it hold the junk data when SafeBuffer is allocated.
It will cause CTS test "InstallDexMetadataHostTest#testProfileSnapshotAfterInstall"
fail if the zero contents is not enabled in libc memory allocator.

Change-Id: I762176c4e5e09cfa7f8dcdfe68df962735c100bd
Test: run cts -m CtsDexMetadataHostTestCases