Do not run background verification on unknown class loaders.
There can be races between the Java threads and the background
verification thread when verifying classes. The Java threads are allowed
to load classes while verifying, but the background verification thread
cannot. Which means we can be in situations where the background
verification sees, for example, a concrete j.l.Integer but an unresolved
j.l.Number.

Test: 833-background-verification
Bug: 202463489
Change-Id: I5aadb9d5ee79003b3d8200571cecd0819e13bd64