Refactor JavaHprofDataSource::Trace() callback lambda
JavaHprofDataSource::Trace() used to have a ~200 lines lambda callback
(with nested lambda). This is an attempt to split into different methods
in order to improve readability by:

* Creating some methods with slightly more meaningful names.
* Adding comments to the methods.
* Limiting the scope of variables: they can be class members, passed
  around as parameters or be local variables.

The only functional change is that `min_nonnull_ptr` used to include
objects that were later set to nullptr. Now the code first filters the
field references (and sets them to nullptr) and later takes the minimum.

Change-Id: I638d4c359dab66aa20ec63122bee05a478699c6b