Dump NativeAllocationRegistry size
Before this commit, perfetto_hprof used to dump objects, types and
references between objects.

This commit introduces a way to also dump values of object fields that
are not references (currently limited to long types).

Only a single field is dumped for now:
libcore.util.NativeAllocationRegistry.size. This is useful to understand
the size of certain native allocations.

The list of fields to dump is hardcoded and very limited. We don't
expect it to be expanded much in the near future.

Bug: 194395971
Change-Id: I8de37fe7b3b6f68ae85a6bbd4c6bba4d0da23df6