Update run-test script to pass --Xjitthreshold:0 for jvmti tests
To provide better test coverage for jvmti + jit, lower the
jit threshold to 0 so we jit on the first use of the function for jvmti
tests. This means that we won't run jvmti tests with the default jit
threshold, but that is OK to omit that case for jvmti tests. We do this
to limit the increase in the runtime of the tests.

Test: art/testrunner.py
Change-Id: Ida0edfccd3ae3f4990f6e863f1db00c0a210803b