Allow extra env uses for StringBuilderAppend optimization
Previously, we wouldn't perform the StringBuilderAppend optimization if
we had extra environment uses. This meant that, for example, calling
substr on the new StringBuilder would block the optimization.

There is a case (also commented with a TODO in the code) in which we do
not eliminate the NewInstance instruction, and that instruction is not
needed in the end either. It is a shortcoming of the current
implementation, but it is still better than previously since we are
removing several Invokes (which also entail e.g. ParallelMove
instructions).

Test: ART tests
Change-Id: I38d7bd7eab8ed6372c73ef3f1a595c94def1aaa7