[Sim] Add a "restricted" setup for the Simulator.
This CL introduces a special ART setup for all modes (host, target),
which is gonna be a base for VIXL Simulator Stage 1 setup. The mode:
  - Limits the GC to only non-concurent Mark&Sweep.
  - Turns off the Nterp (so only switch C++ intepreter to be used).
  - Turns off the JNI compiler (GenericJNITrampoline to be used).
  - Rejects the compilation of all method (except for the allow list).
  - Turns off compilation of intrinsics.

With these limitations it will be easier to start off an initial
Simulator Stage 1 setup. As the limitations are set for all the modes,
it will be easy to compare/debug the workflow for the simulator mode
and to compare it to host and target runs.

The CL also adds sections in knownfailures.json for tests that fail
in this special setup.

Test: test-art-host, test-art-target.

Change-Id: I1bf4843f74b515f3c495951212f21abe166de987