Support calling entry / exit hooks from JIT code for non-debuggable
This CL extends the support to call entry / exit hooks directly
from JITed code for non-debuggable mode. This is done as a best effort
and we won't call method exit hooks for the methods that are already
on the stack when method tracing is enabled. This means it might
be less precise than the existing approach.

The idea is to basically invalidate all the JITed code when we
need method entry / exit hooks and re-JIT them with the support
to call entry / exit hooks as and when necessary.

Test: art/testrunner.py
Change-Id: I79e4547368691a21b26226508228d890369823e9