Update apex search in host_bcp.sh .
Do the apex translation for all BCP components from apexes.
If there is an exact match for the apex directory, do not
search for any other apex names, otherwise look for a ".*"
suffix. We add the "." to avoid confusing apex names such
as "com.android.media" and "com.android.mediaprovider".

Test: host_bcp.sh $ANDROID_PRODUCT_OUT/system/framework/oat/<arch>/services.odex/services.odex
Change-Id: I54b26f4fb23a0aee99ae055cfe376f0c8ea04a79