Exclude run-tests with a custom `run` script from automated executions.
We do not have proper support for ART run-tests with a custom `run`
script (yet). Despite this limitation, we were still including some of
these tests in automated executions. As they might not be running as
intended, disable them until we can support them properly.

This change also adds a curated list of exceptions for some tests with
a custom `run` script that are known to work fine with the default
test execution strategy (i.e. when ignoring their `run` script), even
if not exactly as they would with the original ART run-test harness.

Test: atest --test-mapping art:presubmit
Bug: 147812905
Change-Id: I1e32eb19b68c321e22c4c966ff8db672c4d13560