OpenJDK 11: Merging in java.lang.Integer (toString(int) and helpers)
This is part of merging upstream changes from OpenJDK 11.28. This CL
updates test expectations for 912-classes and 988-moethod-trace
as toString(int) and its helpers methods were updated.

Added char version of getChars() (in addition to byte) not to break
existing and not yet upgraded to 11 code from java.lang.Long and
java.lang.AbstractStringBuilder.

Bug: 207841289
Test: art/test/testrunner/testrunner.py --host -b \
      -t 912-classes -t 988-method-trace
Change-Id: Icc15f60e7aae353dc6212a3352edabb1aa916c1b