Check oat files of boot images.
Before this change, oat files of boot images were not checked when we
load boot images. We need this check to ensure that the boot images are
up-to-date, especially when we move the primary boot image out of the
ART APEX.

This check is performed at an early stage, rather than when the oat
files are actually loaded for execution, in order to determine whether
the runtime should re-compile the boot image or not.

On boot, this check costs 30ms on average for each instruction set
(tested for 10 times on a high-end device), which means a total of 60ms
boot time. However, this cost is inevitable in order to ensure
correctness.

Bug: 209963183
Test: All ART gtests and run-tests in Presubmits.
Change-Id: I5c6f3edd633fd980c51c00d02b813659996e85dd