Relax GC in clear soft references in Zygote
See b/204785359

We are observing a high number of dirty pages caused by removeStaleEntries
in LocaleObjectCache. Tim was able to help nail this down to GC's exception
behavior in Zygote, and as Hans suggested, this leads to reallocating and
reclaiming the locales in Zygote and finishing the cleanup in the children,
which is exactly what's happening causing the COW pages.

And we expect this change, which cleans up SoftReference less aggressively
in Zygote, should have no major negative impact.

How SoftReference could be better handled will need to be further discussed.

Change-Id: I871ffa917e954a5f1f7f86a774b03ced57a1e9d7