Add reference processor tests
Adds a general reference processor test (2042), that tests mostly for
the order in which References are enqueued.

Also adds another test (2043) to check that finalizer reachable data is
preserved to run the finalizer, and that WeakReferences behave as
expected. This one is primarily intended to be (ab)used as a benchmark
to measure Reference.get() times.

Bug: 190867430
Bug: 189738006
Bug: 211784084

Test: Passes on host, Treehugger

Change-Id: I08ddc7c8154ff123345fc1196e224879d0cd92f1