MethodHandles: change asType() / invoke() implementation
This CL changes how we deal with non-exact invokes and asType(). It
drops the concept of `nominalType` from our implementation and adds an
AsTypeAdapter transformer for performing pairwise conversions. It also
switches to use the asTypeCache to reduce the cost of non-exact
invokes.

There still remains some type adaptation code in the native method
handle code, but the majority of this can be removed-up in a follow-up
change.

Bug: 207844518
Test: art/test.py --host
Test: atest CtsLibcoreTestCases

Change-Id: I2faf260b164767104c7ae0bc061fbf4f6579bbff