DO NOT SUBMIT Fix some clang-tidy warnings.
This removes some of the more numerous warnings, reducing the count
from 39k to 27k. Only one warning, performance-no-automatic-move, is
removed completely so it can be converted to an error.

Also try disabling ART cpplint, to see if it fixes the checkbuild
errors with the NOLINT markers that conflict with cpplint.

Test: m tidy-art
Bug: 213953102
Change-Id: I35f8cdd0c7a90a97c908be6b2fb34f840ddc0382