Revert^2 "Add bss support for inlining BCP DexFiles for single image"
This reverts commit 239c449f10853374a9a7a2bd1d8413354054573e.

Reason for revert: Fix in android-review.googlesource.com/c/1968661/1..3

Additionally, we are adding the runtime support for updating the .bss slots on
android-review.googlesource.com/c/1968661/3..4. Since BCP the DexFiles will
not have a corresponding OatDexFile, we can do a best-effort try to fetch the
outer method's OatFile to update the bss slots at runtime.

Bug: 154012332
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: Compile and open top 10 APKs with csuite locally
Change-Id: Ifb2b152cee363c4674fc9273690c73f0406282ba