Remove `ArtField::GetAccessFlagsDCheck()`.
This check provided little value and retrieved the declaring
class with a read barrier which may be undesirable in some
contexts. Also, do not require holding the mutator lock for
`ArtField::GetAccessFlags()` as we do not require it in
`ArtMethod::GetAccessFlags()` either.

We removed a similar check from `ArtMethod` in
    https://android-review.googlesource.com/718816
but that one was even less useful because it was checking an
earlier class state (`kIdx`) than here (`kLoaded`).

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Change-Id: I82fad9a0eb5cc21f67998ecf722facc1d8f8f76c