Add missing android.system package to art-bootclasspath-fragment
The split_packages and package_prefixes properties in
art-bootclasspath-fragment cover all packages provided by that fragment
except "android.system". This change adds "android.system" to make them
define an exhaustive set of packages. That is necessary as an
accompanying change will fail the build if a bootclasspath_fragment
includes a package that is not in that set.

The build check failed on coverage builds due to the presence of extra
jacoco, emma and asm classes so it was necessary to add coverage
specific versions of the properties to art-bootclasspath-fragment.

Bug: 194063708
Test: atest signature_patterns_test
      m out/soong/hiddenapi/hiddenapi-flags.csv
      # Breaks without this change to add android.system to
      # the art-bootclasspath-fragment.
Change-Id: If1bb7083d1137e7a3c7a6eec2706127ac6ca5ff7