Check for java debuggble when using deoptimize flag for instrumentation
ShouldDeoptimizeFlag is used both for CHA and for instrumentation. When
instrumenting the stack, we can use the ShouldDeoptimizeFlag only when
we are in java debuggable. For non java debuggable we still need to
instrument the return address.

This also causes problems when we toggle java debuggable after runtime
has started which will be fixed in a follow up CL.

Bug: 222479430
Test: art/test.py
Change-Id: Id6990d967c79d661a79243861869d615c0a37003