Inliner will return trye when identifying a method as always throws
The reasoning is that we want to run other phases (e.g. dead code
elimination) if we have extra information (i.e. if we now know that
the method always throws).

If we successfully analized a method, we don't encode that in the
return type since we use the return type for control flow with the
meaning of 'we inlined the method'.

Bug: 227316307
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: Ic91463b89e7ccf326416f32b9df5c07489bf48d8