Convert run-test bash build scripts to python.
Tests use the build script for two distinct purposes:

 * Some tests generate their source code before completion.
   Keep this code in base for now and rename it "generate-sources".

 * To customize arguments passed to the default compilation script.
   Since the default compilation script is in python, make this
   a normal python call, and add any customization as arguments.

This speeds up the run-test compilation.  There was small cost to
pay due to the switching from python to bash to python.
This added up for the thousands of run-test compilations that we do.

It also removes the need for argument parser in the default script.
We could also remove the data passing via environment (to-do later).
It moves the definition to more build-system-like look and feel.

Test: The outputs are bit-for-bit identical as before.
Change-Id: Idca4181a4676036f06aae0a8f6eea3a3c30d390e