Add additional checks for ArtMethod::GetOatQuickMethodHeader
We expect a non zero pc for all cases apart from runtime methods
and proxy invoke handlers. Add a DCHECK to assert this case.

Also assert that we would always find AOT code for non-native
methods on the fallback path. This requires fixing a unit test
that was passing a pc value of 0 to get the method header of
existing entry point.

Test: art/test.py
Change-Id: I10cfa84c5b9019227d278dd511a4ea09f6aff5b2