Remove slow tests to prevent LUCI timeouts on slow targets
This CL does two things: shards large tests that were previously run as
a whole package; and removes the slowest tests. It serves as a cheap
measure to prevent LUCI timeouts on slow targets, and should be
addressed better in a separate CL (tracked in http://b/208639267).

TestNg tests run time is down to 11s due to slowest tests removed,
which should impact LUCI much less than before this CL.

Bug: 230738282
Bug: 208639267
Test: art/tools/run-libcore-tests.sh --mode=host
Change-Id: I0545690d28d1178b02e21005bbb555089ba3f856