Don't disable JIT GC when the stack is instrumented
Earlier we needed to disable GC when there are instrumentation stubs or
the stack is instrumented due to some races in updating the entry
points. This is fixed now and we no longer need to disable JIT GC. The
disabling of the JIT GC wasn't accurate too, we were missing some places
where we don't disable GC (for ex: when instrumenting after the process
has started) which was also causing some DCHECK failures. This CL
removes the DCHECK and also updates it to not disable JIT GC even when
we have instrumenented code.

Bug: 231686121
Test: art/test.py
Change-Id: I16bae31c34587f55889769f575c59ec9ed833e0b