Separate kRuntimeISA/kRuntimeQuickCodeISA
Currently ART has an implicit assumption that the ISA of the quick
code (e.g: app code) that is targeted to be run by ART will be the
same as the machine that runs ART itself (e.g: runtime code). This
assumption is made by having a single value (kRuntimeISA)
representing both of these types of code.

Instead kRuntimeISA is split into two values so that it is possible
to target the case where these two ISAs don't match, for example: in
simulator mode. The two values are as follows.
- kRuntimeISA: Must match the ISA of the machine that ART will be
               run on. This ISA will be used for the native context,
               native stack frame and native ABI.
- kRuntimeQuickCodeISA: The ISA that ART will generate quick code
                        for, i.e: that java code will be compiled
                        to. This ISA will be used for the quick
                        context, quick stack frame and quick ABI.

Therefore replace kRuntimeISA with kRuntimeQuickCodeISA where the
semantics of quick code ISA is expected, not the ISA of the machine
the code is running on.

Authors: Artem Serov <artem.serov@linaro.org>
         Chris Jones <christopher.jones@arm.com>

Test: test.py --host --target
Change-Id: Ibe6f3e27179f3baa49a8de89686341df0dcf16f5