Use thread-local test for references
WORK IN PROGRESS. Does not compile.

We used to use a thread local to check whether weak reference access
was enabled with the CC collector, and a static field in the Reference
class otherwise. The static field code really needed an acquire
load to make the test correct, but we never implemented that, and
it might have slowed down code. Instead switch to always using the
thread-local for the test. There we can avoid the acquire load for
everything past the initial load following a GC.

In the process, we removed many different component-specific
weak reference notification mechanisms. There were alsoadditional
follow-on effects, since we previously did some system weak processing
wildly differently in the two cases. Notable secondary effects include:

The monitor code behaves differently for CMS, since that has the
unfortunate property that IsMarked is not always usable for newly
allocated objects. We added an explicit query to determine whether
the current collector needs this handling, rather than tying this
to the weak reference access handling.

We removed the weak reference processing notification mechanism
from system_weak.h.

JVMTI was relying on the lock from that notification mechanism
for other purposes as well. We switched it to a new lock.

Bug: 232459100
Bug: 211784084
Test: broken
Change-Id: I1df8e48b42a6f8a802c6918ecf64c8d5cd5f9328